+++ lib/security/index.js
@@ -1,0 +1,8 @@
+
+"use strict";
+
+module.exports = {
+  BasicAuthSecurity: require('./BasicAuthSecurity')
+, ClientSSLSecurity: require('./ClientSSLSecurity')
+, WSSecurity: require('./WSSecurity')
+};
