+++ lib/utils.js
@@ -1,0 +1,11 @@
+
+"use strict";
+var crypto = require('crypto');
+exports.passwordDigest = function passwordDigest(nonce, created, password) {
+  // digest = base64 ( sha1 ( nonce + created + password ) )
+  var pwHash = crypto.createHash('sha1');
+  var rawNonce = new Buffer(nonce || '', 'base64').toString('binary');
+  pwHash.update(rawNonce + created + password);
+  return pwHash.digest('base64');
+};
+
